+++ src/main/java/DistanceBolt.java
@@ -11,29 +11,29 @@ import java.util.Map;
 
 public class DistanceBolt extends BaseRichBolt {
     OutputCollector _collector;
-    Map<Integer, Integer[]> test;
+    Map<Integer, Double[]> test;
 
     @Override
     public void prepare(Map<String, Object> topoConf, TopologyContext context, OutputCollector collector) {
         _collector = collector;
-        test = new HashMap<Integer, Integer[]>();
+        test = new HashMap<Integer, Double[]>();
     }
 
     public void execute(Tuple input) {
         int id = input.getInteger(0);
-        int x = input.getInteger(1);
-        int y = input.getInteger(2);
+        double x = input.getDouble(1);
+        double y = input.getDouble(2);
 
-        int distance = 0;
+        double distance = 0;
 
         if (test.containsKey(id)) {
-            int lastX = test.get(id)[0];
-            int lastY = test.get(id)[1];
+            double lastX = test.get(id)[0];
+            double lastY = test.get(id)[1];
 
             distance = test.get(id)[2] + Math.abs(lastX - x) + Math.abs(lastY - y);
         }
 
-        test.put(id, new Integer[]{x, y, distance});
+        test.put(id, new Double[]{x, y, distance});
 
         _collector.emit(new Values(id, distance));
     }
